projects/assistant-ia/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic-settings==2.9.1
python-dotenv==1.0.1
asyncpg==0.30.0
apscheduler==3.11.0
pytz==2025.2
